Reduce caching to source image digest only.
After moving to libvips, reading dimensions from an image is as fast as
reading them from a text file, so there is no need to cache them.

- Combine cache classes & modules into one
- Remove all cache related code from GeneratedImage
- Remove most cache-related code from SourceImage
- Remove several unnecessary tests
- Fix broken tests
